Update language on forms/actions & tasks
Updated language on forms/actions to reflect language in new framework overview deck. We're no longer using the term Actions as in "Actions are basically forms." We're talking about using dynamic forms to enable care guides for decision support, surveys, and contact creation/edit forms. See in particular slides 64-69 of the framework overview at the bottom of the CHT Features page: http://communityhealthtoolkit.org/features/

I also extended the description of Tasks to explain why they are generated on-the-fly using only docs that are available on the users' device.